+++ tests/autoarm/test_integration.py
@@ -1,6 +1,7 @@
 import asyncio
 import json
 
+from homeassistant.components.alarm_control_panel.const import ATTR_CHANGED_BY, AlarmControlPanelState
 from homeassistant.components.notify.const import ATTR_TITLE
 from homeassistant.const import ATTR_ICON, CONF_CONDITIONS, CONF_DELAY_TIME
 from homeassistant.core import HomeAssistant
@@ -102,6 +103,24 @@ async def test_broken_condition_raises_issue(hass: HomeAssistant, issue_registry
     }
 
     assert issue.severity == ir.IssueSeverity.ERROR
+
+
+async def test_on_panel_change_ignores_autoarm_generated_event(hass: HomeAssistant) -> None:
+
+    assert await async_setup_component(hass, "autoarm", CONFIG)
+    await hass.async_block_till_done()
+
+    hass.states.async_set("binary_sensor.button_middle", "on")
+    await hass.async_block_till_done()
+    assert hass.states.get("autoarm.last_intervention").state == "button"  # type: ignore
+    panel_entity = hass.states.get("alarm_panel.testing")
+    assert panel_entity is not None
+    assert panel_entity.attributes[ATTR_CHANGED_BY] == "autoarm.button"
+
+    # when alarm panel is changed directly, this is recorded as an intervention
+    hass.states.async_set("alarm_panel.testing", "armed_vacation")
+    await hass.async_block_till_done()
+    assert hass.states.get("autoarm.last_intervention").state == "alarm_panel"  # type: ignore
 
 
 async def test_arm_on_away(hass: HomeAssistant) -> None: